Support for running inside Expo
When using Expo, you can't load native modules with ejecting which kinda ruins the Expo experience.

Fortunately, Expo ships with it's own LinearGradient component.

With this commit, the LinearGradient will be loaded from Expo if Expo is installed, making it possible to use this module out-of-box in an Expo environment